Fixed find navigator bindings and made text color the control accent color when something else is selected.
